Refactored profiles.py removing base directory logic from it by
moving it qt/aqt/__init__.py and set most OpenGl settings before
creating the first application.
